Move data processing to own function.
Uncoupling the parsing of data from the InputDataHandler class into it's own separate function allows us to do error checking before we coin an instance of the InputDataHandler class. There's no point create this class if data errors are found as this errors will be presented to the user to fix first.